Build 313; development of the GameTimer;

+++ DX11Framework/ImageReaders/ImageReader.h
@@ -29,5 +29,5 @@ public:
 	bool ReadRawImageData(const std::string & filePath,
 		UINT & imageWidth,
 		UINT & imageHeight,
-		_Inout_ std::vector<float> & imageData);
+		_Inout_ std::vector<uint8_t> & imageData);
 };
